chore: better plugin structure
- Avoid global dependencies.
- Use proper WPML API.
- Better feature copy.

Ref: wpmlbridge-238
