+++ src/Feature.php
@@ -3,40 +3,56 @@
 namespace WPML\ElasticPress;
 
 class Feature extends \ElasticPress\Feature {
-	/** @var LanguageSearch */
-	private $languageSearch;
+	/** @var Field\Search */
+	private $fieldSearch;
 
-	/** @var IndexingLangParam */
-	private $indexLangParam;
+	/** @var Field\Sync */
+	private $fieldSync;
+
+	/** @var Sync\Dashboard */
+	private $syncDashboard;
+
+	/** @var Sync\CLI */
+	private $syncCli;
 
 	/**
-	 * @param  LanguageSearch  $languageSearch
+	 * @param Field\Search   $fieldSearch
+	 * @param Field\Sync     $fieldSync
+	 * @param Sync\Dashboard $syncDashboard
+	 * @param Sync\CLI       $syncCli
 	 */
-	public function __construct( LanguageSearch $languageSearch, IndexingLangParam $indexLangParam ) {
-		parent::__construct();
-
-		$this->languageSearch = $languageSearch;
-		$this->indexLangParam = $indexLangParam;
+	public function __construct(
+		Field\Search   $fieldSearch,
+		Field\Sync     $fieldSync,
+		Sync\Dashboard $syncDashboard,
+		Sync\CLI       $syncCli
+	) {
+		$this->fieldSearch   = $fieldSearch;
+		$this->fieldSync     = $fieldSync;
+		$this->syncDashboard = $syncDashboard;
+		$this->syncCli       = $syncCli;
 
 		$this->slug                     = 'wpml';
-		$this->title                    = __( 'Integration WPML with ElasticPress', 'sitepress' );
+		$this->title                    = esc_html__( 'WPML integration', 'sitepress' );
 		$this->requires_install_reindex = false;
+
+		parent::__construct();
 	}
 
 	public function setup() {
-		add_filter( 'ep_post_sync_args_post_prepare_meta', [ $this->languageSearch, 'addLangInfo' ], 10, 2 );
-		add_filter( 'ep_post_formatted_args', [ $this->languageSearch, 'filterByLang' ], 10, 1 );
-
-		$this->indexLangParam->addHooks();
+		$this->fieldSearch->addHooks();
+		$this->fieldSync->addHooks();
+		$this->syncDashboard->addHooks();
+		$this->syncCli->addHooks();
 	}
 
 	public function output_feature_box_summary() {
-		$content = esc_html__( 'Integration WPML with ElasticPress', 'sitepress' );
+		$content = esc_html__( 'Index and search content in its specific language.', 'sitepress' );
 		echo '<p>' . $content . '</p>';
 	}
 
 	public function output_feature_box_long() {
-		$content = esc_html__( 'This allows to search for content in a specific language only.', 'sitepress' );
+		$content = esc_html__( 'Index your content with the right stopwords, and get search results in the relevant frontend language.', 'sitepress' );
 		echo '<p>' . $content . '</p>';
 	}
